requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
beautifulsoup4>=4.10.0
selenium>=4.0.0
requests>=2.26.0
pydantic>=1.8.2
pytest>=6.2.5
python-dotenv>=0.19.0
webdriver-manager>=3.5.2
pandas>=1.3.0  # For CSV export functionality 